requirements.txt
Django==2.2.4
django-cors-headers==3.1.0
django-redis==4.10.0
django-rest-registration==0.5.0
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0
mysqlclient==1.4.4
Pillow==6.1.0
PyJWT==1.7.1
pytz==2019.2
redis==3.3.7
sqlparse==0.3.0